[Chore] Move modals from pages/dashboard/modals to components/modalContents
