+++ src/pages/dashboard/[dashboardId]/index.tsx
@@ -4,12 +4,12 @@ import Column from "@/components/common/Column";
 import Header from "@/components/common/Header";
 import { SkeletonColumn } from "@/components/common/Skeleton";
 import { createColumn, updateColumn, deleteColumn } from "@/api/column.api";
-import CreateCardModal from "@/pages/dashboard/modals/CreateCardModal";
-import ManageColumnModal from "@/pages/dashboard/modals/ManageColumnModal";
-import AddColumnModal from "@/pages/dashboard/modals/AddColumnModal";
+import CreateCardModal from "@/components/ModalContents/CreateCardModal";
+import ManageColumnModal from "@/components/ModalContents/ManageColumnModal";
+import AddColumnModal from "@/components/ModalContents/AddColumnModal";
 import useCardForm from "@/hooks/useCardForm";
 import useColumnForm from "@/hooks/useColumnForm";
-import EditCardModal from "@/pages/dashboard/modals/EditCardModal";
+import EditCardModal from "@/components/ModalContents/EditCardModal";
 import { useFetchColumns } from "@/hooks/useFetchColumns";
 import useEditCardForm from "@/hooks/useEditCard";
 import { useEditCardSubmit } from "@/hooks/useEditCardSubmit";
